+++ family_foto/__init__.py
@@ -7,8 +7,8 @@ from flask_admin import Admin
 from flask_debugtoolbar import DebugToolbarExtension
 from flask_login import LoginManager
 
-from family_foto.admin.AdminIndexView import AdminHomeView
-from family_foto.admin.AdminModelView import AdminModelView
+from family_foto.admin.admin_index_view import AdminHomeView
+from family_foto.admin.admin_model_view import AdminModelView
 from family_foto.const import UPLOADED_PHOTOS_DEST_RELATIVE, UPLOADED_VIDEOS_DEST_RELATIVE, \
     RESIZED_DEST_RELATIVE, RESIZED_DEST
 from family_foto.logger import log
