+++ family_foto/admin/admin_model_view.py
